Update DST for RFC 24 and fix dispatch bug
